Scripts: Fix issues introduced in the initial commit
The original project was uploaded from a cygwin environment, thus
the access mode of the script files are not correct.
This patch fixes the access mode of the script files and update
the README-win32.txt documentation accordingly along with some
indentation fixes.
This patch also adds a missing .gitignore control file for the
src/bug2issue module.

Signed-off-by: Lv Zheng <lv.zheng@intel.com>